Loop scraper forever until shutdown
- Add a time metric to get metrics on how long this takes
- Add a .no_shutdown breaker to disable shutdowns

+++ python/scraper/ec2_runscript.sh
@@ -10,20 +10,41 @@ aws s3 cp s3://relliker.com/scraper/relliker-settings.source relliker-settings.s
 source relliker-settings.source
 aws s3 cp s3://relliker.com/scraper/group_scraper-1.0.0.tar.gz group_scraper-1.0.0.tar.gz
 pip install group_scraper-1.0.0.tar.gz
-python -m group_scraper
-if [ $? == 0 ]; then
-  aws cloudwatch put-metric-data --region=us-west-2 \
-      --metric-name "Success" \
-      --namespace "Relliker/Scraper"\
-      --value 1 \
-      --timestamp $(date --utc +%FT%T.%3NZ)
-else
-  aws cloudwatch put-metric-data --region=us-west-2 \
-      --metric-name "Success" \
-      --namespace "Relliker/Scraper"\
-      --value 0 \
-      --timestamp $(date --utc +%FT%T.%3NZ)
-fi
-popd
 
-sudo shutdown -h now
+function reportExitCode {
+  aws cloudwatch put-metric-data --region=us-west-2 \
+      --metric-name "Success" \
+      --namespace "Relliker/Scraper"\
+      --value $1 \
+      --timestamp $(date --utc +%FT%T.%3NZ)
+}
+
+function reportRuntime {
+  aws cloudwatch put-metric-data --region=us-west-2 \
+      --metric-name "Time" \
+      --namespace "Relliker/Scraper"\
+      --value $1 \
+      --unit "Seconds" \
+      --timestamp $(date --utc +%FT%T.%3NZ)
+}
+
+EXIT_CODE=0
+while [ "$EXIT_CODE" == 0 ]; do
+  START=$(date +%s)
+  python -m group_scraper
+  EXIT_CODE=$?
+  echo "Exit code: $EXIT_CODE"
+  END=$(date +%s)
+  RUNTIME=$((END-START))
+  if [ "$EXIT_CODE" == 0 ]; then
+    reportExitCode 1
+  else
+    reportExitCode 0
+  fi
+  reportRuntime $RUNTIME
+done
+
+popd
+if [ ! -e ".no_shutdown" ]; then
+  sudo shutdown -h now
+fi
